% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{dic.svdraws}
\alias{dic.svdraws}
\title{Deviance information criterion for stochastic volatility models}
\usage{
dic.svdraws(x)
}
\description{
Deviance information criterion for stochastic volatility models
}
\examples{
# Example 1
## Simulate a short and highly persistent SV process 
sim <- svsim(100, mu = -10, phi = 0.99, sigma = 0.2)

## Obtain 5000 draws from the sampler (that's not a lot)
draws <- svsample(sim, draws = 5000, burnin = 100,
		  priormu = c(-10, 1), priorphi = c(20, 1.5), priorsigma = 0.2)

## Calculate DIC
dic.svdraws(draws)
}
\keyword{dic}
