% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{data}
\name{get_default_fast_sv}
\alias{get_default_fast_sv}
\alias{get_default_general_sv}
\alias{default_fast_sv}
\title{Default Values for the Expert Settings}
\format{
An object of class \code{list} of length 11.
}
\usage{
get_default_fast_sv()

get_default_general_sv(priorspec)

default_fast_sv
}
\arguments{
\item{priorspec}{a \code{priorspec} object created by
\code{\link{specify_priors}}}
}
\description{
These functions define meaningful expert settings for
argument \code{expert} of \code{\link{svsample}} and its derivatives.
The result of \code{get_default_fast_sv} should be provided as
\code{expert$fast_sv} and \code{get_default_general_sv} as \code{expert$general_sv}
when relevant.
}
\seealso{
\code{link{svsample}}, \code{\link{specify_priors}}, \code{\link{svsample_roll}}, \code{\link{svsample_fast_cpp}}, \code{\link{svsample_general_cpp}}
}
\keyword{datasets}
