% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{paratraceplot}
\alias{paratraceplot}
\title{Trace Plot of MCMC Draws from the Parameter Posteriors}
\usage{
paratraceplot(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Called for its side effects. Returns argument \code{x} invisibly.
}
\description{
Generic function for plotting iterations vs. sampled parameter values.
A detailed help for the method implemented in \pkg{stochvol} can be found in
\code{\link{paratraceplot.svdraws}}.
}
\seealso{
Other plotting: \code{\link{paradensplot}},
  \code{\link{paratraceplot.svdraws}},
  \code{\link{plot.svdraws}}, \code{\link{plot.svpredict}},
  \code{\link{volplot}}
}
\concept{plotting}
\keyword{hplot}
