% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{svlsample}
\alias{svlsample}
\title{Main function}
\usage{
svlsample(y, draws = 10000, burnin = 1000, designmatrix = NA,
  priormu = c(0, 100), priorphi = c(5, 1.5), priorsigma = 1,
  priorrho = c(3, 5), priorbeta = c(0, 10000), thinpara = 1,
  thinlatent = 1, thintime = 1, quiet = FALSE, startpara,
  startlatent, expert, ...)
}
\description{
Some text about \code{svlsample}.
}
\examples{

## Simulate a short and highly persistent SV process
sim <- svsim(100, mu = -10, phi = 0.99, sigma = 0.2, rho = -0.4)

## Obtain 5000 draws from the sampler (that's not a lot)
draws <- svlsample(sim$y, draws = 2000, burnin = 500,
		  priormu = c(-10, 1), priorphi = c(20, 1.5), priorsigma = 0.2, priorrho = c(3, 6))

## Check out the results
summary(draws)
plot(draws)

}
