% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{svlsample2}
\alias{svlsample2}
\title{Minimal overhead version of \code{\link{svlsample}}.}
\usage{
svlsample2(y, draws = 1, burnin = 0, priormu = c(0, 100),
  priorphi = c(5, 1.5), priorsigma = 1, priorrho = c(3, 5),
  thinpara = 1, thinlatent = 1, thintime = 1, quiet = FALSE,
  startpara, startlatent)
}
\description{
\code{svsample2} is a minimal overhead version of \code{\link{svsample}}
with slightly different default arguments and a simplified return value
structure. It is intended to be used mainly for one-step updates where speed
is an issue, e.g., as a plug-in into other MCMC samplers. Note that
absolutely no input checking is performed, thus this function is to be used
with proper care!
}
