% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_and_process_expert}
\alias{validate_and_process_expert}
\title{Validate and Process Argument 'expert'}
\usage{
validate_and_process_expert(expert = NULL, priorspec = specify_priors())
}
\arguments{
\item{expert}{list, the input values for expert.}

\item{priorspec}{a \code{priorspec} object created by
\code{\link{specify_priors}}}
}
\value{
A list that is the input extended by default values. If
the input is invalid, an error is thrown.
}
\description{
A helper function that validates the input and extends it with
default values if there are missing parts for argument 'expert'.
}
\seealso{
\code{\link{specify_priors}}
}
\concept{validation}
