% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochvol-package.R, R/utilities_svdraws.R
\name{extractors}
\alias{extractors}
\alias{para}
\alias{latent}
\alias{latent0}
\alias{priors}
\alias{thinning}
\alias{runtime}
\alias{sampled_parameters}
\alias{vola}
\title{Common Extractors for 'svdraws' Objects}
\usage{
para(x)

latent(x)

latent0(x)

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

vola(x)
}
\arguments{
\item{x}{\code{svdraws} object.}
}
\value{
The return value depends on the actual funtion:
\item{para(x)}{extracts the parameter draws and returns them as an
\code{mcmc} object.}
\item{latent(x)}{extracts the latent contemporaneous
log-volatility draws and returns them as an \code{mcmc} object.}
\item{latent0(x)}{extracts the latent initial log-volatility draws and
returns as an \code{mcmc} object.}
\item{priors(x)}{extracts the prior
parameters used and returns them in a \code{list}.}
\item{thinning(x)}{extracts the thinning parameters used and returns them in
a \code{list}.}
\item{runtime(x)}{extracts the runtime and returns it as a
\code{proc_time} object.}
}
\description{
Some simple extractors returning the corresponding element of an
\code{svdraws} object.
}
\keyword{utilities}
