% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{logret}
\alias{logret}
\alias{logret.default}
\title{Computes the log returns of a time series}
\usage{
logret(dat, demean = FALSE, standardize = FALSE, ...)

\method{logret}{default}(dat, demean = FALSE, standardize = FALSE, ...)
}
\arguments{
\item{dat}{The raw data.}

\item{demean}{Logical value indicating whether the data should
be de-meaned.}

\item{standardize}{Logical value indicating whether the data should
be standardized (in the sense that each component series has an empirical
variance equal to one).}

\item{...}{Ignored.}
}
\value{
Log returns of the (de-meaned / standardized) data.
}
\description{
\code{logret} computes the log returns of a time
series, with optional de-meaning and/or standardization.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Log returns of vectors
}}

\concept{utilities}
