% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochvol-cpp-doc.R
\name{update_regressors}
\alias{update_regressors}
\title{Single MCMC update of Bayesian linear regression}
\usage{
update_regressors(dependent_variable, independent_variables, beta, prior_spec)
}
\arguments{
\item{dependent_variable}{the left hand side}

\item{independent_variables}{the matrix of the independent variables. Has to be of same height as the length of the dependent variable}

\item{beta}{the vector of the latent states used in MDA. Updated in place}

\item{prior_spec}{prior specification object. See type_definitions.h}
}
\description{
Samples the coefficients of a linear regression. The prior
distribution is multivariate normal and it is specified in
prior_spec.
}
\seealso{
Other stochvol_cpp: 
\code{\link{update_fast_sv}()},
\code{\link{update_general_sv}()},
\code{\link{update_t_error}()}
}
\concept{stochvol_cpp}
\keyword{update}
