% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochvol-cpp-doc.R
\name{update_t_error}
\alias{update_t_error}
\title{Single MCMC Update of Student's t-Distribution}
\usage{
update_t_error(homosked_data, tau, nu, prior_spec)
}
\arguments{
\item{homosked_data}{de-meaned and homoskedastic observations}

\item{tau}{the vector of the latent states used in MDA. Updated in place}

\item{nu}{parameter nu. The degrees of freedom for the t-distribution. Updated in place}

\item{prior_spec}{prior specification object. See type_definitions.h}
}
\description{
Samples the degrees of freedom parameter of de-meaned and homoskedastic
t-distributed input variates. Marginal data augmentation (MDA) is applied, tau
is the vector of auxiliary latent states.
Depending on the prior specification, nu might not be updated, just tau.
}
\seealso{
Other stochvol_cpp: 
\code{\link{update_fast_sv}()},
\code{\link{update_general_sv}()},
\code{\link{update_regressors}()}
}
\concept{stochvol_cpp}
\keyword{update}
